feat-refactor: Subscription
- Implemented subscription feature
- Implemented subscription APIs
- Implemented code refactor

+++ src/main/java/com/socialmedia/service/UserService.java
@@ -24,7 +24,7 @@ public class UserService {
     }
 
     public List<String> getUserStack(User user) {
-        return userRepository.findById(user.getId().intValue())
+        return userRepository.findById(user.getId())
                 .map(User::getStack)
                 .orElse(Collections.emptyList());
     }
